IgnoredCommentTombstone renders for replies too
